pyproject.toml
[tool.poetry]
name = "live-wallpaper"
version = "0.1.0"
description = "live wallpaper"
authors = ["so1n <qaz6803609@163.com>"]
license = "Apache Software License"
readme = "./README.md"
repository = "https://github.com/so1n/live-wallpaper"
homepage = "https://github.com/so1n/live-wallpaper"

[tool.poetry.dependencies]
python = "^3.7"
croniter = "^1.0.15"
python-crontab = "^2.5.1"
requests = "^2.26.0"
pydantic = "^1.8.2"
Pillow = "^8.3.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.15.0"
isort = "^5.9.3"
black = "^21.8b0"
mypy = "^0.910"
flake8 = "^3.9.2"
autoflake = "^1.4"
types-requests = "^2.25.6"
types-croniter = "^1.0.0"

[tool.black]
line-length = 120
target-version = ['py37']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120

[tool.pytest]
testpaths = "tests"

[tool.coverage.run]
branch = true
source = ["rap"]

[tool.coverage.report]
exclude_lines = [
    # Have to re-enable the standard pragma
    "pragma: no cover",

    # Don't complain about missing debug-only code:
    "def __repr__",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # type hints
    "if TYPE_CHECKING:",
    "pass"
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
